requirements.txt
ffmpeg-python==0.2.0
openai==0.27.10
langchain==0.0.321
python-dotenv==1.0.0
transformers==4.34.0
torch==2.0.1
Pillow==10.0.0
